+++ src/ast.rs
@@ -28,6 +28,7 @@ pub enum Expression {
     Sum(Branch, MaybeLevel),
     /// `split { Bla x => y }`
     Split(Branch),
+    Merge(Box<Self>, Box<Self>),
     /// `\Pi a: b. c`
     Pi(Typed, Box<Self>, MaybeLevel),
     /// `\Sigma a: b. c`
@@ -85,10 +86,13 @@ pub type GenericBranch<T> = BTreeMap<String, Box<T>>;
 /// Pattern matching branch.
 pub type Branch = GenericBranch<Expression>;
 
-pub fn branch_to_righted(branch: Branch) -> GenericBranch<Either<Value, Expression>> {
-    let mut case_tree: GenericBranch<Either<Value, Expression>> = Default::default();
+/// This function name is mysterious, but I failed to find a better name. It's for converting a
+/// `Branch` into a `CaseTree` by inserting the `context` to every `Case`s.
+pub fn branch_to_righted(branch: Branch, context: Telescope) -> CaseTree {
+    let mut case_tree: CaseTree = Default::default();
     for (name, expression) in branch.into_iter() {
-        case_tree.insert(name, Box::new(Either::Right(*expression)));
+        let case = GenericCase::new(Either::Right(*expression), context.clone());
+        case_tree.insert(name, Box::new(case));
     }
     case_tree
 }
@@ -138,7 +142,10 @@ pub enum GenericNeutral<Value: Clone> {
     /// Neutral form: stuck on trying to find the second element of a free variable.
     Second(Box<Self>),
     /// Neutral form: stuck on trying to case-split a free variable.
-    Split(GenericCaseTree<Either<Value, Expression>, Value>, Box<Self>),
+    Split(
+        GenericBranch<GenericCase<Either<Value, Expression>, Value>>,
+        Box<Self>,
+    ),
 }
 
 /// `Neut` in Mini-TT, neutral value.
@@ -156,13 +163,6 @@ pub enum Pattern {
     Var(String),
 }
 
-/// Whether a type is recursive or not.
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub enum DeclarationType {
-    Simple,
-    Recursive,
-}
-
 /// `Decl` in Mini-TT.
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct Declaration {
@@ -170,7 +170,7 @@ pub struct Declaration {
     pub prefix_parameters: Vec<Typed>,
     pub signature: Expression,
     pub body: Expression,
-    pub declaration_type: DeclarationType,
+    pub is_recursive: bool,
 }
 
 impl Declaration {
@@ -180,14 +180,14 @@ impl Declaration {
         prefix_parameters: Vec<Typed>,
         signature: Expression,
         body: Expression,
-        declaration_type: DeclarationType,
+        is_recursive: bool,
     ) -> Self {
         Self {
             pattern,
             prefix_parameters,
             signature,
             body,
-            declaration_type,
+            is_recursive,
         }
     }
 
@@ -198,13 +198,7 @@ impl Declaration {
         signature: Expression,
         body: Expression,
     ) -> Self {
-        Self::new(
-            pattern,
-            prefix_parameters,
-            signature,
-            body,
-            DeclarationType::Simple,
-        )
+        Self::new(pattern, prefix_parameters, signature, body, false)
     }
 
     /// Recursive declarations
@@ -214,13 +208,7 @@ impl Declaration {
         signature: Expression,
         body: Expression,
     ) -> Self {
-        Self::new(
-            pattern,
-            prefix_parameters,
-            signature,
-            body,
-            DeclarationType::Recursive,
-        )
+        Self::new(pattern, prefix_parameters, signature, body, true)
     }
 }
 
@@ -305,36 +293,33 @@ pub enum Closure {
 
 /// Generic definition for three kinds of case trees
 #[derive(Debug, Clone, Eq, PartialEq)]
-pub struct GenericCaseTree<ValueExpression, ValueInScope: Clone> {
-    pub branches: Box<GenericBranch<ValueExpression>>,
-    pub environment: Box<TelescopeRc<ValueInScope>>,
+pub struct GenericCase<Expression, Value: Clone> {
+    pub expression: Expression,
+    pub context: TelescopeRc<Value>,
 }
 
-impl<Expr, Value: Clone> GenericCaseTree<Expr, Value> {
-    pub fn new(branches: Box<GenericBranch<Expr>>, environment: Box<TelescopeRc<Value>>) -> Self {
+impl<Expression, Value: Clone> GenericCase<Expression, Value> {
+    pub fn new(expression: Expression, context: TelescopeRc<Value>) -> Self {
         Self {
-            branches,
-            environment,
+            expression,
+            context,
         }
     }
-
-    pub fn boxing(branches: GenericBranch<Expr>, environment: TelescopeRc<Value>) -> Self {
-        Self::new(Box::new(branches), Box::new(environment))
-    }
-
-    pub fn destruct(self) -> (GenericBranch<Expr>, TelescopeRc<Value>) {
-        let GenericCaseTree {
-            branches,
-            environment,
-        } = self;
-        (*branches, *environment)
-    }
 }
+
+/// One single case in case trees.
+pub type Case = GenericCase<Either<Value, Expression>, Value>;
 
 /// `SClos` in Mini-TT.<br/>
 /// Case tree.
-pub type CaseTree = GenericCaseTree<Either<Value, Expression>, Value>;
+pub type CaseTree = GenericBranch<Case>;
 
-pub fn reduce_to_value(either: Either<Value, Expression>, context: Telescope) -> Value {
-    either.either(|l| l, |r| r.eval(context))
+impl GenericCase<Either<Value, Expression>, Value> {
+    pub fn reduce_to_value(self) -> Value {
+        let GenericCase {
+            expression,
+            context,
+        } = self;
+        expression.either(|l| l, |r| r.eval(context))
+    }
 }