+++ src/check/expr.rs
@@ -1,13 +1,13 @@
-use crate::ast::{
-    reduce_to_value, up_var_rc, Branch, Closure, Expression, GenericCaseTree, Pattern, Value,
-};
+use std::collections::BTreeMap;
+
+use either::Either;
+
+use crate::ast::MaybeLevel::SomeLevel;
+use crate::ast::{up_var_rc, Branch, Closure, Expression, GenericCase, Pattern, Value};
 use crate::check::decl::check_declaration;
 use crate::check::read_back::generate_value;
 use crate::check::subtype::check_subtype;
 use crate::check::tcm::{update_gamma, update_gamma_borrow, TCE, TCM, TCS};
-use either::Either;
-use std::collections::BTreeMap;
-use crate::ast::MaybeLevel::SomeLevel;
 
 /// `checkI` in Mini-TT.<br/>
 /// Type inference rule. More inferences are added here (maybe it's useful?).
@@ -26,8 +26,8 @@ pub fn check_infer(index: u32, mut tcs: TCS, expression: Expression) -> TCM<Valu
             let mut map = BTreeMap::new();
             let context = tcs.context.clone();
             let inferred = Either::Left(check_infer(index, tcs, *expression)?);
-            map.insert(name, Box::new(inferred));
-            Ok(Value::Sum(GenericCaseTree::boxing(map, context), 0))
+            map.insert(name, Box::new(GenericCase::new(inferred, context)));
+            Ok(Value::Sum(map, 0))
         }
         Pair(left, right) => {
             let left = check_infer(index, tcs_borrow!(tcs), *left)?;
@@ -104,7 +104,7 @@ pub fn check_type(index: u32, tcs: TCS, expression: Expression) -> TCM<(u32, TCS
             let inferred = check_infer(index, tcs_borrow!(tcs), expression)?;
             match inferred.level_safe() {
                 SomeLevel(level) if level > 0 => Ok((level - 1, tcs)),
-                _ => Err(TCE::NotTypeType(inferred))
+                _ => Err(TCE::NotTypeType(inferred)),
             }
         }
     }
@@ -149,13 +149,12 @@ pub fn check(index: u32, mut tcs: TCS, expression: Expression, value: Value) -> 
             )
         }
         (E::Constructor(name, body), V::Sum(constructors, _)) => {
-            let constructor = *constructors
-                .branches
+            let constructor = constructors
                 .get(&name)
                 .ok_or_else(|| TCE::InvalidConstructor(name))?
-                .clone();
-            let constructor_type = reduce_to_value(constructor, *constructors.environment);
-            check(index, tcs, *body, constructor_type)
+                .clone()
+                .reduce_to_value();
+            check(index, tcs, *body, constructor)
         }
         (E::Sum(constructors, _), V::Type(level)) => {
             check_level(level, check_sum_type(index, tcs, constructors)?)
@@ -180,14 +179,14 @@ pub fn check(index: u32, mut tcs: TCS, expression: Expression, value: Value) -> 
         }
         // I really wish to have box pattern here :(
         (E::Split(mut branches), V::Pi(sum, closure, _)) => match *sum {
-            V::Sum(sum_branches, _) => {
-                for (name, branch) in sum_branches.branches.into_iter() {
+            V::Sum(sum_branches, level) => {
+                for (name, branch) in sum_branches.into_iter() {
                     let pattern_match = match branches.remove(&name) {
                         Some(pattern_match) => *pattern_match,
                         None => return Err(TCE::MissingCase(name)),
                     };
                     let signature = V::Pi(
-                        Box::new(reduce_to_value(*branch, *sum_branches.environment.clone())),
+                        Box::new(branch.reduce_to_value()),
                         Closure::Choice(Box::new(closure.clone()), name.clone()),
                         0,
                     );